+++ http/src/main/scala/com/twitter/finatra/http/modules/MustacheModule.scala
@@ -2,10 +2,10 @@ package com.twitter.finatra.http.modules
 
 import com.github.mustachejava.{DefaultMustacheFactory, Mustache, MustacheFactory}
 import com.google.inject.Provides
-import com.twitter.finatra.annotations.Flag
 import com.twitter.finatra.http.internal.marshalling.mustache.ScalaObjectHandler
 import com.twitter.finatra.http.routing.FileResolver
 import com.twitter.inject.TwitterModule
+import com.twitter.inject.annotations.Flag
 import java.io._
 import javax.inject.Singleton