+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/doeverything/main/services/ComplexService.scala
@@ -1,7 +1,7 @@
 package com.twitter.finatra.http.tests.integration.doeverything.main.services
 
 import com.google.inject.assistedinject.Assisted
-import com.twitter.finatra.annotations.Flag
+import com.twitter.inject.annotations.Flag
 import javax.inject.{Inject, Named}
 import org.joda.time.Duration