+++ inject/inject-app/src/main/java/com/twitter/inject/annotations/Flag.java
@@ -1,4 +1,4 @@
-package com.twitter.finatra.annotations;
+package com.twitter.inject.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;